(PC-33579) refactor(CategoryButton): remove v1
