+++ src/features/search/components/CategoriesButtonsDisplay/CategoriesButtonsDisplay.tsx
@@ -3,22 +3,32 @@ import { Dimensions, FlatList, ListRenderItem, Platform, View } from 'react-nati
 import styled, { useTheme } from 'styled-components/native'
 
 import { VenueMapLocationModal } from 'features/location/components/VenueMapLocationModal'
-import {
-  CategoryButton,
-  CategoryButtonProps,
-} from 'features/search/components/CategoryButton/CategoryButton'
+import { Gradient } from 'features/search/enums'
 import { VenueMapBlock } from 'features/venueMap/components/VenueMapBlock/VenueMapBlock'
 import { useShouldDisplayVenueMap } from 'features/venueMap/hook/useShouldDisplayVenueMap'
-import { useFeatureFlag } from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
-import { RemoteStoreFeatureFlags } from 'libs/firebase/firestore/types'
 import { LocationMode } from 'libs/location/types'
 import { getMediaQueryFromDimensions } from 'libs/react-responsive/useMediaQuery'
-import { CategoryButtonV2 } from 'shared/Buttons/CategoryButtonV2'
+import { CategoryButton } from 'shared/Buttons/CategoryButton'
 import { theme } from 'theme'
 import { useModal } from 'ui/components/modals/useModal'
+import { AccessibleIcon } from 'ui/svg/icons/types'
 import { Spacer, TypoDS, getSpacing } from 'ui/theme'
+// eslint-disable-next-line no-restricted-imports
+import { ColorsEnum } from 'ui/theme/colors'
 import { getHeadingAttrs } from 'ui/theme/typographyAttrs/getHeadingAttrs'
 
+export type CategoryButtonProps = {
+  label: string
+  Illustration?: FunctionComponent<AccessibleIcon>
+  baseColor?: ColorsEnum
+  gradients: Gradient
+  onPress: () => void
+  children?: never
+  // v2 App Design
+  textColor: ColorsEnum
+  fillColor: ColorsEnum
+  borderColor: ColorsEnum
+}
 export type ListCategoryButtonProps = CategoryButtonProps[]
 
 type Props = {
@@ -26,17 +36,11 @@ type Props = {
   children?: never
 }
 
-const CategoryButtonItem: ListRenderItem<CategoryButtonProps> = ({ item }) => (
-  <CategoryButtonContainer>
-    <CategoryButton {...item} />
-  </CategoryButtonContainer>
-)
-
 const CATEGORY_BUTTON_HEIGHT_SEARCH = getSpacing(24.25)
 
-const CategoryButtonItemV2: ListRenderItem<CategoryButtonProps> = ({ item }) => (
+const CategoryButtonItem: ListRenderItem<CategoryButtonProps> = ({ item }) => (
   <CategoryButtonContainer>
-    <CategoryButtonV2 {...item} height={CATEGORY_BUTTON_HEIGHT_SEARCH} />
+    <CategoryButton {...item} height={CATEGORY_BUTTON_HEIGHT_SEARCH} />
   </CategoryButtonContainer>
 )
 
@@ -44,9 +48,6 @@ const isWeb = Platform.OS === 'web'
 
 export const CategoriesButtonsDisplay: FunctionComponent<Props> = ({ sortedCategories }) => {
   const { shouldDisplayVenueMap, selectedLocationMode } = useShouldDisplayVenueMap()
-  const enableNewCategoryBlock = useFeatureFlag(
-    RemoteStoreFeatureFlags.WIP_APP_V2_SEARCH_CATEGORY_BLOCK
-  )
 
   const {
     showModal: showVenueMapLocationModal,
@@ -64,7 +65,7 @@ export const CategoriesButtonsDisplay: FunctionComponent<Props> = ({ sortedCateg
   return (
     <FlatList
       data={sortedCategories}
-      renderItem={enableNewCategoryBlock ? CategoryButtonItemV2 : CategoryButtonItem}
+      renderItem={CategoryButtonItem}
       keyExtractor={(item) => item.label}
       numColumns={numColumns}
       key={numColumns} // update key to avoid the following error: Changing numColumns on the fly is not supported. Change the key prop on FlatList when changing the number of columns to force a fresh render of the component.
@@ -81,7 +82,7 @@ export const CategoriesButtonsDisplay: FunctionComponent<Props> = ({ sortedCateg
             </React.Fragment>
           ) : null}
 
-          {enableNewCategoryBlock ? <CategoriesTitleV2 /> : <CategoriesTitle />}
+          <CategoriesTitleV2 />
           <VenueMapLocationModal
             visible={venueMapLocationModalVisible}
             dismissModal={hideVenueMapLocationModal}
@@ -95,14 +96,6 @@ export const CategoriesButtonsDisplay: FunctionComponent<Props> = ({ sortedCateg
   )
 }
 
-const CategoriesTitle = styled(TypoDS.Title3).attrs({
-  children: 'Explore les catégories',
-  ...getHeadingAttrs(2),
-})(({ theme }) => ({
-  marginTop: getSpacing(4),
-  marginBottom: getSpacing(theme.isDesktopViewport ? 1 : 2),
-  paddingHorizontal: getSpacing(2),
-}))
 const CategoriesTitleV2 = styled(TypoDS.Title4).attrs({
   children: 'Parcours les catégories',
   ...getHeadingAttrs(2),
